# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Documentation for ITA-Wiki",
        "description": "Project ITA-Wiki documentation wall",
        "version": "1.0.0.0"
    },
    "paths": {
...
        "/api/roles-node": {
            "put": {
                "tags": [
                    "RolesNode"
                ],
                "summary": "Update an existing role using node_id",
                "description": "Allows an authorized user to update the role of a specific GitHub node_id.",
                "operationId": "21da04f149f343df852087700cd96884",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "node_id",
                                    "role",
                                    "authorized_node_id"
                                ],
                                "properties": {
                                    "node_id": {
                                        "description": "GitHub node_id of the user to update the role",
                                        "type": "string",
                                        "example": "MDQ6VXNlcjY3Mjk2MDg="
                                    },
                                    "role": {
                                        "description": "Role to be assigned",
                                        "type": "string",
                                        "enum": [
                                            "superadmin",
                                            "admin",
                                            "mentor",
                                            "student"
                                        ],
                                        "example": "mentor"
                                    },
                                    "authorized_node_id": {
                                        "description": "GitHub node_id of the user making the request (must have permissions, e.g. superadmin)",
                                        "type": "string",
                                        "example": "MDQ6VXNlcjE="
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Cannot update a role of user of equal or higher role rank or make an update to an equal or higher rank than your own",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You cannot update a role that is equal or higher than your own."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The request contains an invalid role."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "RolesNode"
                ],
                "summary": "Create a new role using node_id",
                "description": "Allows an authorized user to create a new role for a specific GitHub node_id.\n     *\n     *     Note: The 'node_id' field must be unique. If you receive a 'has already been taken' error, please use a different value for 'node_id'.",
                "operationId": "059334212f6a3e439277b3d6281800fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "node_id",
                                    "role",
                                    "authorized_node_id"
                                ],
                                "properties": {
                                    "node_id": {
                                        "description": "GitHub node_id of the user to assign the role. Must be unique and not already exist in the database.",
                                        "type": "string",
                                        "example": "MDQ6VXNlcjUNIQUE="
                                    },
                                    "role": {
                                        "description": "Role to be assigned",
                                        "type": "string",
                                        "enum": [
                                            "superadmin",
                                            "admin",
                                            "mentor",
                                            "student"
                                        ],
                                        "example": "mentor"
                                    },
                                    "authorized_node_id": {
                                        "description": "GitHub node_id of the user making the request (must have permissions, e.g. superadmin)",
                                        "type": "string",
                                        "example": "MDQ6VXNlcjE="
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Cannot create a role equal or higher than your own",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You cannot create a role equal or higher than your own."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "authorized_node_id": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The selected authorized node id is invalid."
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The request contains an invalid role."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login-node": {
            "post": {
                "tags": [
                    "RolesNode"
                ],
                "summary": "Retrieve a role by GitHub node_id",
                "description": "Fetches a role using the provided GitHub node_id. If the role does not exist, it returns an error.",
                "operationId": "65a22473b9a9ff9e2be253682bb520ca",
                "parameters": [
                    {
                        "name": "node_id",
                        "in": "query",
                        "description": "GitHub node_id of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "MDQ6VXNlcjY3Mjk2MDg="
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role found."
                                        },
                                        "role": {
                                            "$ref": "#/components/schemas/RoleNode"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role not found."
                                        },
                                        "role": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/feature-flags/role-self-assignment-node": {
            "put": {
                "tags": [
                    "RolesNode"
                ],
                "summary": "Role Self Assignment Node",
                "description": "Updates a role using the provided GitHub node_id. If the role does not exist, it returns an error.",
                "operationId": "5b8285921a6c2150c4d6fe312e21c325",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "node_id",
                                    "role"
                                ],
                                "properties": {
                                    "node_id": {
                                        "type": "string",
                                        "example": "MDQ6VXNlcjY3Mjk2MDg="
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El Rol se ha actualizado."
                                        },
                                        "role": {
                                            "properties": {
                                                "node_id": {
                                                    "type": "string",
                                                    "example": "MDQ6VXNlcjY3Mjk2MDg="
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: The feature flag for role self-assignment is disabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La autoasignación de roles ha sido desactivada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La petición contiene un node_id inexistente en nuestro sistema."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleNode": {
                "title": "RoleNode",
                "description": "RoleNode object representing a user's role and associated GitHub node_id",
                "properties": {
                    "node_id": {
                        "description": "The GitHub node_id of the user",
                        "type": "string",
                        "example": "MDQ6VXNlcjY3Mjk2MDg="
                    },
                    "role": {
                        "description": "The role of the user",
                        "type": "string",
                        "example": "student"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-17T19:23:41.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-17T19:23:41.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}